requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
aiosqlite==0.19.0
apscheduler==3.10.4
python-dotenv==1.0.0
sgp4==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
